requirements.txt
ib_insync
pandas
llama-index
llama-index-llms-huggingface
jinja2
streamlit
python-dotenv
langchain_experimental
sentence-transformers
langchain
langchain_community
langchainhub
huggingface_hub==0.34.1
transformers
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-llms-huggingface-api
"llama-index[llms-huggingface]"